DCO Remediation Commit for Aksha Thakkar <thaakb@amazon.com>
I, Aksha Thakkar <thaakb@amazon.com>, hereby add my Signed-off-by to this commit: c8cbaae11a39e953a43406bb21f5c97dc0b9bbee
